# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BeatSight Backend",
    "description": "\n## BeatSight Web API\n\nBeatSight is an AI-powered drum beatmap generation platform. This API provides:\n\n### Features\n- **Authentication**: JWT-based auth with access/refresh tokens\n- **Songs**: Upload, manage, and organize your music library\n- **AI Jobs**: Queue songs for AI beatmap generation with real-time progress\n- **Cloud Sync**: Sync preferences and progress across devices\n- **Storage**: Secure file upload with presigned URLs\n\n### Authentication\nMost endpoints require authentication. Include the JWT token in the Authorization header:\n```\nAuthorization: Bearer <access_token>\n```\n\n### Rate Limits\n- Standard users: 100 requests/minute\n- Premium users: 500 requests/minute\n- AI job quota: Based on subscription tier\n    ",
    "contact": {
      "name": "BeatSight Support",
      "url": "https://beatsight.app/support",
      "email": "support@beatsight.app"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/storage/audio/{song_id}": {
      "post": {
        "tags": [
          "storage"
        ],
        "summary": "Upload audio file",
        "description": "Upload an audio file for a song.\n\nAccepts MP3, WAV, or FLAC audio files up to 100MB.",
        "operationId": "upload_audio_api_storage_audio__song_id__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Song Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_audio_api_storage_audio__song_id__post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "Download audio file",
        "description": "Download the original audio file for a song.\n\nReturns the audio as a streaming response.",
        "operationId": "download_audio_api_storage_audio__song_id__get",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Song Id"
            }
          },
          {
            "name": "extension",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Audio format",
              "default": "mp3",
              "title": "Extension"
            },
            "description": "Audio format"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/storage/audio/{song_id}/url": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "Get presigned URL for audio download",
        "description": "Get a presigned URL for direct audio download.\n\nThe URL can be used for client-side playback without authentication.",
        "operationId": "get_audio_download_url_api_storage_audio__song_id__url_get",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Song Id"
            }
          },
          {
            "name": "extension",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Audio format",
              "default": "mp3",
              "title": "Extension"
            },
            "description": "Audio format"
          },
          {
            "name": "expires_in",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 86400,
              "minimum": 60,
              "default": 3600,
              "title": "Expires In"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/storage/audio/{song_id}/upload-url": {
      "post": {
        "tags": [
          "storage"
        ],
        "summary": "Get presigned URL for audio upload",
        "description": "Get a presigned URL for direct audio upload.\n\nUse this for large files to upload directly to storage without\nproxying through the API server.",
        "operationId": "get_audio_upload_url_api_storage_audio__song_id__upload_url_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Song Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignedUrlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/storage/stems/{song_id}/{stem}": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "Download stem file",
        "description": "Download a separated stem file.\n\nAvailable stems: drums, bass, vocals, other",
        "operationId": "download_stem_api_storage_stems__song_id___stem__get",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Song Id"
            }
          },
          {
            "name": "stem",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Stem name (drums, bass, vocals, other)",
              "title": "Stem"
            },
            "description": "Stem name (drums, bass, vocals, other)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/storage/stems/{song_id}/{stem}/url": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "Get presigned URL for stem download",
        "description": "Get a presigned URL for direct stem download.",
        "operationId": "get_stem_download_url_api_storage_stems__song_id___stem__url_get",
        "parameters": [
          {
            "name": "song_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Song Id"
            }
          },
          {
            "name": "stem",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Stem"
            }
          },
          {
            "name": "expires_in",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 86400,
              "minimum": 60,
              "default": 3600,
              "title": "Expires In"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/storage/beatmaps/{map_id}/v{version}": {
      "post": {
        "tags": [
          "storage"
        ],
        "summary": "Upload beatmap file",
        "description": "Upload a beatmap file for a specific version.\n\nThe beatmap should be in BeatSight JSON format (.bs).",
        "operationId": "upload_beatmap_api_storage_beatmaps__map_id__v_version__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Map Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_beatmap_api_storage_beatmaps__map_id__v_version__post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "Download beatmap file",
        "description": "Download a beatmap file.\n\nReturns the beatmap JSON content.",
        "operationId": "download_beatmap_api_storage_beatmaps__map_id__v_version__get",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Map Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/storage/beatmaps/{map_id}/v{version}/url": {
      "get": {
        "tags": [
          "storage"
        ],
        "summary": "Get presigned URL for beatmap download",
        "description": "Get a presigned URL for direct beatmap download.",
        "operationId": "get_beatmap_download_url_api_storage_beatmaps__map_id__v_version__url_get",
        "parameters": [
          {
            "name": "map_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Map Id"
            }
          },
          {
            "name": "version",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version"
            }
          },
          {
            "name": "expires_in",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 86400,
              "minimum": 60,
              "default": 3600,
              "title": "Expires In"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignedUrlResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_audio_api_storage_audio__song_id__post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "Audio file (MP3, WAV, FLAC)"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_audio_api_storage_audio__song_id__post"
      },
      "Body_upload_beatmap_api_storage_beatmaps__map_id__v_version__post": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "Beatmap JSON file"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_beatmap_api_storage_beatmaps__map_id__v_version__post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PresignedUrlRequest": {
        "properties": {
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "description": "Content type for upload (required for PUT URLs)",
            "default": "audio/mpeg"
          },
          "expires_in": {
            "type": "integer",
            "maximum": 86400.0,
            "minimum": 60.0,
            "title": "Expires In",
            "description": "URL validity in seconds (1 min to 24 hours)",
            "default": 3600
          }
        },
        "type": "object",
        "title": "PresignedUrlRequest",
        "description": "Request for presigned URL generation."
      },
      "PresignedUrlResponse": {
        "properties": {
          "url": {
            "type": "string",
            "title": "Url",
            "description": "Presigned URL for direct access"
          },
          "expires_at": {
            "type": "string",
            "title": "Expires At",
            "description": "ISO timestamp when URL expires"
          },
          "method": {
            "type": "string",
            "title": "Method",
            "description": "HTTP method to use (GET or PUT)"
          }
        },
        "type": "object",
        "required": [
          "url",
          "expires_at",
          "method"
        ],
        "title": "PresignedUrlResponse",
        "description": "Response with presigned URL."
      },
...
      "UploadResponse": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key",
            "description": "Storage key for the uploaded file"
          },
          "size": {
            "type": "integer",
            "title": "Size",
            "description": "File size in bytes"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type",
            "description": "MIME type of the file"
          }
        },
        "type": "object",
        "required": [
          "key",
          "size",
          "content_type"
        ],
        "title": "UploadResponse",
        "description": "Response after successful upload."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}